franquise-api/requirements.txt
Django==3.2.11
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
